feat: Combine configs using deepmerge
Prior to this change, all top-level keys of a config file would be
replaced entirely by the incoming config. For some apps, this may
be sufficient. However, some apps have complex settings, and the
extend feature is much more useful if you can be more granular about
what you are overriding.

With this change, you can keep as much or as little of the base config
as you'd like.

closes #21